feat: implement check exist user and show message
