+++ src/Components/Header/Header.jsx
@@ -1,10 +1,85 @@
-import { Navbar, Nav, NavDropdown, Button, Container, Modal, Form } from 'react-bootstrap';
+import { Navbar, Nav, NavDropdown, Container, Modal } from 'react-bootstrap';
 import { useState } from 'react';
 import authService from "../../services/authService"
 import usersService from "../../services/usersService"
 import styles from './Header.module.scss'
+import { withRouter } from 'react-router-dom'
 
-const Header = () => {
+import { useFormik } from 'formik';
+import * as yup from 'yup';
+import Button from '@material-ui/core/Button';
+import TextField from '@material-ui/core/TextField';
+
+
+const validationSchemaRegister = yup.object().shape({
+    name: yup
+        .string('Введите имя')
+        .required('Необходимо ввести имя'),
+    email: yup
+        .string('Введите e-mail')
+        .email('E-mail должен быть вида: email@mark.com')
+        .required('Необходимо ввести e-mail'),
+    password: yup
+        .string('Введите пароль')
+        .min(8, 'Пароль должен быть не короче 8 символов и содержать хотябы одну заглавную букву')
+        .required('Необходимо ввести пароль'),
+});
+
+const validationSchemaLogIn = yup.object().shape({
+    email: yup
+        .string('Введите e-mail')
+        .email('E-mail должен быть вида: email@mark.com')
+        .required('Необходимо ввести e-mail'),
+    password: yup
+        .string('Введите пароль')
+        .min(8, 'Пароль должен быть не короче 8 символов и содержать хотябы одну заглавную букву')
+        .required('Необходимо ввести пароль'),
+});
+
+
+const Header = ({ history }) => {
+
+    const formikRegistr = useFormik({
+        initialValues: {
+            name: '',
+            email: '',
+            password: '',
+        },
+        validationSchema: validationSchemaRegister,
+        onSubmit: async (values) => {
+            const user = {
+                name: values.name,
+                email: values.email,
+                password: values.password,
+            }
+
+            const data = await usersService.register(user)
+            if (data.data) {
+                await authService.login(user.email, user.password);
+                handleCloseRegister();
+                setLogged(authService.getJwt());
+            } else {
+                handleAddRegisterError()
+            }
+        }
+    });
+
+    const formikLogIn = useFormik({
+        initialValues: {
+            email: '',
+            password: '',
+        },
+        validationSchema: validationSchemaLogIn,
+        onSubmit: async (values) => {
+            const data = await authService.login(values.email, values.password) || {};
+            if (data.message) {
+                handleCloseLogIn();
+                setLogged(authService.getJwt())
+            } else {
+                handleAddLogInError()
+            }
+        }
+    });
 
     const [showRegister, setShowRegister] = useState(false);
     const handleCloseRegister = () => setShowRegister(false);
@@ -14,46 +89,34 @@ const Header = () => {
     const handleCloseLogIn = () => setShowLogIn(false);
     const handleShowLogIn = () => setShowLogIn(true);
 
-    const [email, setEmail] = useState('');
-    const [password, setPassword] = useState('');
-    const [name, setName] = useState('');
+    const [showRegisterError, setShowRegisterError] = useState(false);
+    const handleDeleteRegisterError = () => setShowRegisterError(false);
+    const handleAddRegisterError = () => setShowRegisterError(true);
+
+    const [showLogInError, setShowLogInError] = useState(false);
+    const handleDeleteLogInError = () => setShowLogInError(false);
+    const handleAddLogInError = () => setShowLogInError(true);
 
     const [isLogged, setLogged] = useState(authService.getJwt())
 
-
-    const handleChangeName = (event) => {
-        setName(event.target.value);
-    }
-
-    const handleChangeEmailLogIn = (event) => {
-        setEmail(event.target.value);
-    }
-
-    const handleChangePasswordLogIn = (event) => {
-        setPassword(event.target.value);
-    }
-
-    async function handleSubmitRegister() {
-        const user = {
-            name: name,
-            email: email,
-            password: password,
-        }
-        await usersService.register(user);
-        await authService.login(user.email, user.password);
-        handleCloseRegister();
-        setLogged(authService.getJwt())
-    }
-
-    async function handleSubmitLogIn() {
-        await authService.login(email, password);
-        handleCloseLogIn();
-        setLogged(authService.getJwt())
-    }
-
-    function handleLogOut(e) {
+    function handleLogOut() {
         authService.logout();
-        setLogged(null)
+        setLogged(null);
+        history.push('/');
+    }
+
+    let registerErrorMessage;
+    if (showRegisterError) {
+        registerErrorMessage = <div className={styles.errorMessage}>Пользователь с таким e-mail уже существует</div>
+    } else {
+        registerErrorMessage = <div className={styles.errorMessageHidden}>Пользователь с таким e-mail уже существует</div>
+    }
+
+    let logInErrorMessage;
+    if (showLogInError) {
+        logInErrorMessage = <div className={styles.errorMessage}>Пользователя с таким e-mail не существует</div>
+    } else {
+        logInErrorMessage = <div className={styles.errorMessageHidden}>Пользователя с таким e-mail не существует</div>
     }
 
     let buttonRegister;
@@ -101,84 +164,88 @@ const Header = () => {
                 </Container>
             </Navbar>
 
-            <Modal show={showRegister} onHide={handleCloseRegister}>
+            <Modal show={showRegister} onHide={handleCloseRegister} onClick={() => handleDeleteRegisterError()}>
                 <Modal.Header closeButton>
-                    <Modal.Title>Регистрация</Modal.Title>
+                    <Modal.Title>Регистрация
+                        {registerErrorMessage}
+                    </Modal.Title>
                 </Modal.Header>
                 <Modal.Body>
-                    <Form onSubmit={handleSubmitRegister}>
-                        <Form.Group controlId="formBasicName">
-                            <Form.Label>Имя пользователя</Form.Label>
-                            <Form.Control
-                                type="text"
-                                placeholder="Имя пользователя"
-                                value={name}
-                                onChange={handleChangeName}
-                            />
-                        </Form.Group>
-                        <Form.Group controlId="formBasicEmail">
-                            <Form.Label>Email</Form.Label>
-                            <Form.Control
-                                type="text"
-                                placeholder="E-mail"
-                                value={email}
-                                onChange={handleChangeEmailLogIn}
-                            />
-                            <Form.Text className="text-muted">
-                                E-mail должен быть вида: email@mark.com
-                            </Form.Text>
-                        </Form.Group>
-                        <Form.Group controlId="formBasicPassword">
-                            <Form.Label>Пароль</Form.Label>
-                            <Form.Control
-                                type="password"
-                                placeholder="Пароль"
-                                value={password}
-                                onChange={handleChangePasswordLogIn}
-                            />
-                            <Form.Text className="text-muted">
-                                Пароль должен быть не короче 8 символов и содержать хотябы одну заглавную букву
-                            </Form.Text>
-                        </Form.Group>
-                        <Button variant="dark" type="submit">
+                    <form onSubmit={formikRegistr.handleSubmit}>
+                        <TextField
+                            fullWidth
+                            id="name"
+                            name="name"
+                            label="Введите имя"
+                            value={formikRegistr.values.name}
+                            onChange={formikRegistr.handleChange}
+                            error={formikRegistr.touched.name && Boolean(formikRegistr.errors.name)}
+                            helperText={formikRegistr.touched.name && formikRegistr.errors.name}
+                        />
+                        <TextField
+                            fullWidth
+                            id="email"
+                            name="email"
+                            label="Введите Email"
+                            value={formikRegistr.values.email}
+                            onChange={formikRegistr.handleChange}
+                            error={formikRegistr.touched.email && Boolean(formikRegistr.errors.email)}
+                            helperText={formikRegistr.touched.email && formikRegistr.errors.email}
+                        />
+                        <TextField
+                            fullWidth
+                            id="password"
+                            name="password"
+                            label="Введите пароль"
+                            type="password"
+                            value={formikRegistr.values.password}
+                            onChange={formikRegistr.handleChange}
+                            error={formikRegistr.touched.password && Boolean(formikRegistr.errors.password)}
+                            helperText={formikRegistr.touched.password && formikRegistr.errors.password}
+                        />
+                        <Button color="primary" variant="contained" fullWidth type="submit">
                             Зарегистрироваться
-                        </Button>
-                    </Form>
+                    </Button>
+                    </form>
                 </Modal.Body>
             </Modal>
-
-            <Modal show={showLogIn} onHide={handleCloseLogIn} className={styles.modal}>
+            <Modal show={showLogIn} onHide={handleCloseLogIn} className={styles.modal} onClick={() => handleDeleteLogInError() }>
                 <Modal.Header closeButton>
-                    <Modal.Title>Вход</Modal.Title>
+                    <Modal.Title>Вход
+                        {logInErrorMessage}
+                    </Modal.Title>
                 </Modal.Header>
                 <Modal.Body>
-                    <Form onSubmit={handleSubmitLogIn}>
-                        <Form.Group controlId="formBasicEmail">
-                            <Form.Label>Email</Form.Label>
-                            <Form.Control
-                                type="text"
-                                placeholder="Email"
-                                value={email}
-                                onChange={handleChangeEmailLogIn}
-                            />
-                        </Form.Group>
-                        <Form.Group controlId="formBasicPassword">
-                            <Form.Label>Пароль</Form.Label>
-                            <Form.Control
-                                type="password"
-                                placeholder="Пароль"
-                                value={password}
-                                onChange={handleChangePasswordLogIn}
-                            />
-                        </Form.Group>
-                        <Button variant="dark" type="submit">
+                    <form onSubmit={formikLogIn.handleSubmit}>
+                        <TextField
+                            fullWidth
+                            id="email"
+                            name="email"
+                            label="Email"
+                            value={formikLogIn.values.email}
+                            onChange={formikLogIn.handleChange}
+                            error={formikLogIn.touched.email && Boolean(formikLogIn.errors.email)}
+                            helperText={formikLogIn.touched.email && formikLogIn.errors.email}
+                        />
+                        <TextField
+                            fullWidth
+                            id="password"
+                            name="password"
+                            label="Пароль"
+                            type="password"
+                            value={formikLogIn.values.password}
+                            onChange={formikLogIn.handleChange}
+                            error={formikLogIn.touched.password && Boolean(formikLogIn.errors.password)}
+                            helperText={formikLogIn.touched.password && formikLogIn.errors.password}
+                        />
+                        <Button color="primary" variant="contained" fullWidth type="submit">
                             Войти
                     </Button>
-                    </Form>
+                    </form>
                 </Modal.Body>
             </Modal>
         </>
     )
 }
 
-export default Header
+export default withRouter(Header)
