+++ src/Components/Header/Header.module.scss
@@ -1,3 +1,12 @@
 .modal {
     width: 100vw !important;
 }
+
+.errorMessage {
+    font-size: 18px;
+    color: red;
+}
+.errorMessageHidden {
+    font-size: 18px;
+    color: #fff;
+}
