+++ src/services/authService.js
@@ -13,8 +13,10 @@ export function getJwt() {
 
 export async function login(email, password) {  
   const { data } = await http.post(apiEndpoint, { email, password });
-  localStorage.setItem(tokenKey, data.token);
-  localStorage.setItem(userId, data.userId);
+  if (data) {
+    localStorage.setItem(tokenKey, data.token);
+    localStorage.setItem(userId, data.userId);
+  }  
   return data;
 }
